Update CalEvent.java
Added some of the user input prompts.
Concat the values to make it easier to write in a value.
So far I have the summary and location working.
Not sure if this is the order in which it is supposed to go in the file. Does the location matter? Will check this after the rest of it is up and running.