Update CalEvent.java
So I think this is my final file. I've tested it and it works.
So it's able to take two dates and I added the timezone thing. Not sure if we need the standard one too (because of what the check in says). Let me know if you want me to put it in.